Renamed "helper.h" to "util.h".
